Advanced elevator customization
* Select which elevators can be randomized.
* Select possible destinations for *One-way, anywhere*.
* Randomize Sky Temple Gateway, Aerie Access and Aerie elevators.

+++ randovania/game_description/echoes_game_specific.py
@@ -18,7 +18,7 @@ class EchoesBeamConfiguration:
 
 @dataclasses.dataclass(frozen=True)
 class EchoesGameSpecific:
-    energy_per_tank: float
+    energy_per_tank: int
     safe_zone_heal_per_second: float
     beam_configurations: Tuple[EchoesBeamConfiguration, ...]
     dangerous_energy_tank: bool
